refactor(viewer): hoist collection fetches to Workspace, gate project-tab auto-open (VIS-775)
Two related shell simplifications, per PR review (RC2) + my F1 finding:

1. Collection fetches hoisted (RC2): the Library used to fire 12 fetch*
   actions on mount; move them to Workspace.jsx so the route container
   loads the project's objects once. Library becomes a pure consumer
   (via useLibraryData). Other workspace surfaces (canvas, outline,
   project editor) inherit the same load on the route mount.

2. Project tab no longer auto-reopens after the user closes it (F1):
   the hydration effect re-ran on every route change, so closing the
   project tab and navigating made it pop back. A useRef gates the
   auto-open to the first time `project` resolves; after that the
   project tab is user-managed.

- Library.jsx drops the 12 fetch selectors + mount-load useEffect (and
  the now-unused useEffect import).
- Library.test.jsx drops the fetch stubs + the "fetches every collection
  on mount" regression test (no longer applies to the Library).
- Workspace.jsx adds the 12 fetch selectors + mount-load useEffect, and
  the useRef gate around the project-tab open.
- Workspace.test.jsx stubs the 12 fetch actions, adds the
  fetches-on-mount regression test, and a regression test asserting the
  project tab stays closed after user dismissal.

101 workspace tests pass, lint clean.
